dse: More unevaluated expressions
