+++ devito/dse/extended_sympy.py
@@ -3,8 +3,8 @@ Extended SymPy hierarchy.
 """
 
 import sympy
-from sympy import Expr
-from sympy.core.basic import _aresame
+from sympy import Expr, Float, Indexed, S
+from sympy.core.basic import _aresame, Basic
 from sympy.functions.elementary.trigonometric import TrigonometricFunction
 
 
@@ -47,7 +47,7 @@ class taylor_sin(TrigonometricFunction):
 
     @classmethod
     def eval(cls, arg):
-        return 0.0 if arg == 0.0 else eval_taylor_sin(arg)
+        return eval_taylor_sin(arg)
 
 
 class taylor_cos(TrigonometricFunction):
@@ -69,7 +69,7 @@ class bhaskara_sin(TrigonometricFunction):
 
     @classmethod
     def eval(cls, arg):
-        return 0.0 if arg == 0.0 else eval_bhaskara_sin(arg)
+        return eval_bhaskara_sin(arg)
 
 
 class bhaskara_cos(TrigonometricFunction):
@@ -85,13 +85,56 @@ class bhaskara_cos(TrigonometricFunction):
 
 # Utils
 
-def eval_bhaskara_sin(angle):
-    return 16.0*angle*(3.1416-abs(angle))/(49.3483-4.0*abs(angle)*(3.1416-abs(angle)))
+def eval_bhaskara_sin(expr):
+    return 16.0*expr*(3.1416-abs(expr))/(49.3483-4.0*abs(expr)*(3.1416-abs(expr)))
 
 
-def eval_taylor_sin(angle):
-    return angle-(angle*angle*angle/6.0*(1.0-angle*angle/20.0))
+def eval_taylor_sin(expr):
+    v = expr + Mul(-1/6.0,
+                   Mul(expr, expr, expr, evaluate=False),
+                   1.0 + Mul(Mul(expr, expr, evaluate=False), -0.05, evaluate=False),
+                   evaluate=False)
+    try:
+        Float(expr)
+        return v.doit()
+    except TypeError, ValueError:
+        return v
 
 
-def eval_taylor_cos(angle):
-    return 1.0-.5*angle*angle*(1.0-angle*angle/12.0)
+def eval_taylor_cos(expr):
+    expr_square = Mul(expr, expr, evaluate=False)
+    b =  1.0-.5*expr_square*(1.0-expr_square/12.0)
+    v = 1.0 + Mul(-0.5,
+                  Mul(expr, expr, evaluate=False),
+                  1.0 + Mul(expr, expr, -1/12.0, evaluate=False),
+                  evaluate=False)
+    try:
+        Float(expr)
+        return b, v.doit()
+    except TypeError, ValueError:
+        return v
+
+
+def unevaluate_arithmetic(expr):
+    """
+    Reconstruct ``expr`` turning all :class:`sympy.Mul` and :class:`sympy.Add`
+    into, respectively, :class:`devito.Mul` and :class:`devito.Add`.
+    """
+    if expr.is_Float:
+        return expr.func(*expr.atoms())
+    elif isinstance(expr, Indexed):
+        return expr.func(*expr.args)
+    elif expr.is_Symbol:
+        return expr.func(expr.name)
+    elif expr in [S.Zero, S.One, S.NegativeOne, S.Half]:
+        return expr.func()
+    elif expr.is_Atom:
+        return expr.func(*expr.atoms())
+    elif expr.is_Add:
+        rebuilt_args = [unevaluate_arithmetic(e) for e in expr.args]
+        return Add(*rebuilt_args, evaluate=False)
+    elif expr.is_Mul:
+        rebuilt_args = [unevaluate_arithmetic(e) for e in expr.args]
+        return Mul(*rebuilt_args, evaluate=False)
+    else:
+        return expr.func(*[unevaluate_arithmetic(e) for e in expr.args])
